+++ Changelog.md
@@ -1,6 +1,7 @@
 Improved Mobs 1.8.0
 ================
 - Port to 1.18
+- Improved pathfinding of mobs able to break blocks
 - Renamed SWIMMRIDE flag to GUARDIAN
 - Parrots: new flag PARROT.
   Mobs can start riding a parrot to reach normally unreachable places